Add the functions of failure recovery and auto-scaling. For auto-scaling, there are three types: 1. Sub-topology level horizonal scaling up or down. For this type, the entire requested sub-topology can be scaled up or down.
2. VM level horizonal scaling.
For this type, the requested VM can be scaled up. These VM can come from different sub-topologies and the network topology is scaled according to the original VM at the same time.

3. VM level vertical scaling.
For this type, the requested VM can be vertically scaled up or down. Vertical scaling here means that the computing capability of the VM can be scaled. For instance, original VM "nodeA" has 1 vCPU and 1GB memory. According to the request, it can be scaled to 1 vCPU and 3GB memory. Meanwhile, the network connection remains the same.

I also design the infrastructure code for achieving this.

For sub-topology level horizonal scaling, the infrastructure code is simillar as:

- CodeType: "SEQ"
  OpCode:
    Operation: "hscale"
    Options:
        ReqID: "hst1"
        CP: "ExoGENI"
        DC: "GWU (Washington DC,  USA) XO Rack"
        UpDown: UP
        ScaledSTName: null
        STClass: null
        SEClass: null
        VEClass: null
    ObjectType: SubTopology
    Objects: "scaling_1||scaling_2"
- CodeType: "SEQ"
  OpCode:
    Operation: "hscale"
    ObjectType: REQ
    Objects: "hst1"

For VM level horizonal scaling, the infrastructure code is simillar as:

- CodeType: "SEQ"
  OpCode:
    Operation: "hscale"
    Options:
        ReqID: "hvm1"
        CP: "ExoGENI"
        DC: "GWU (Washington DC,  USA) XO Rack"
        ScaledSTName: null
        STClass: null
        SEClass: null
        VEClass: null
    ObjectType: VM
    Objects: "scaling_1.Node0||scaling_2.Node2"
- CodeType: "SEQ"
  OpCode:
    Operation: "hscale"
    ObjectType: REQ
    Objects: "hvm1"

For VM level vertical scaling, the infrastructure code is simillar as:

- CodeType: "SEQ"
  OpCode:
    Operation: "vscale"
    Options:
        ReqID: "v1"
        CPU: 1
        MEM: 3
        STClass: null
        SEClass: null
        VEClass: null
    ObjectType: VM
    Objects: "scaling_1.Node0||scaling_2.Node2"
- CodeType: "SEQ"
  OpCode:
    Operation: "vscale"
    ObjectType: REQ
    Objects: "v1"

All the field with the value of "null" can be omitted.
Currently, all these functions have been test with ExoGENI.